Модуль 3 задание 1 Исправлены замечания, ver.2
